Added window mode delay workaround for mac window size change
Absolute bodge, seems like a race condition - the MacOS display renderer is setting the fullscreen flag outside of a fullscreen transition and this causes it to crash/hang. Slowly debugging the window mode changes with breakpoints seems to allow enough time for this to resolve, so we've added this configurable delay as a workaround for now.